mudanças nos params de valores deposito despesa e abastecidas

+++ src/database/migrations/20220919172045-restocks.js
@@ -34,19 +34,19 @@ module.exports = {
         allowNull: true,
       },
       value_fuel: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: true,
       },
       liters_fuel: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: true,
       },
       total_value_fuel: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: true,
       },
       total_nota_value: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: true,
       },
       created_at: {
