+++ src/database/migrations/20220919172123-deposit_moneys.js
@@ -34,7 +34,7 @@ module.exports = {
         allowNull: true,
       },
       value: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL,
         allowNull: true,
       },
       proof_img: {
